+++ src/Symfony/Bundle/FrameworkBundle/EventListener/SessionListener.php
@@ -12,31 +12,16 @@
 namespace Symfony\Bundle\FrameworkBundle\EventListener;
 
 use Symfony\Component\HttpKernel\EventListener\SessionListener as BaseSessionListener;
-use Symfony\Component\DependencyInjection\ContainerInterface;
+
+@trigger_error(sprintf('The %s class is deprecated since version 3.3 and will be removed in 4.0. Use %s instead.', SessionListener::class, BaseSessionListener::class), E_USER_DEPRECATED);
 
 /**
  * Sets the session in the request.
  *
  * @author Fabien Potencier <fabien@symfony.com>
+ *
+ * @deprecated since version 3.3, to be removed in 4.0. Use {@link BaseSessionListener} instead
  */
 class SessionListener extends BaseSessionListener
 {
-    /**
-     * @var ContainerInterface
-     */
-    private $container;
-
-    public function __construct(ContainerInterface $container)
-    {
-        $this->container = $container;
-    }
-
-    protected function getSession()
-    {
-        if (!$this->container->has('session')) {
-            return;
-        }
-
-        return $this->container->get('session');
-    }
 }